Feat: Switch Air Fan with Hour Button
- 환풍기 아워 버튼으로 끄고 키기 구현

- FanButtonController에서 버튼 눌림 상태 변화 감지
- 눌림 상태 변화가 너무 빨리 반복되면 무시하는 시간 간격을 둬서 연속 입력 무시하도록 처리 (애매하게 밟아서 매우 짧은 시간 안에 다시 눌린 경우를 위함)

- 기존 FullAirFanTrigger 제거
