+++ ClockMate/Assets/Scripts/Desert/Puzzle2/FanButtonController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1e01504905d3cc849a0e14fba965e0c2
+guid: 1cfe0947255dab14d9f51ff80738a39d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
